added tests

+++ src-java/server42/server42-control/src/main/java/org/openkilda/server42/control/kafka/KafkaRecordFilter.java
@@ -35,7 +35,7 @@ public class KafkaRecordFilter implements RecordFilterStrategy<Object, Object> {
 
     private Set<String> switchList;
 
-    private Map<Integer, List<String>> vlanToSwitchMap;
+    private Map<Long, List<String>> vlanToSwitchMap;
 
     public KafkaRecordFilter(@Autowired SwitchToVlanMapping switchToVlanMapping) {
         vlanToSwitchMap = switchToVlanMapping.getVlan();
